+++ CHANGELOG.md
@@ -9,6 +9,11 @@ x.y.z Release notes (yyyy-MM-dd)
 ### Breaking changes
 
 * Combine IDKit setup with PACECloudSDK setup. `IDKit.setup(...)` is no longer accessible. The `IDKit` is now initialized via `PACECloudSDK.setup(...)`. Therefore an optional `OIDConfiguration` parameter was added to the `Configuration` class of the `PACECloudSDK`, which must be initialized with at least the `clientId` and the `redirectUri`.
+* Change properties of `CofuGasStation` from `var` to `val`
+
+### Enhancements
+
+* Add the attribute `properties: Map<String, Any>` to the `CofuGasStation` which includes different information of the gas station e.g. app URL
 
 ### Fixes